+++ bin/generate-flatpak-sources-for-gradle-dependencies.sh
@@ -5,5 +5,6 @@
 cp ./build/flatpak/dependencies-sources.json ./distribution/flatpak/gradle-sources-root.json
 cp ./modules/app-core/build/flatpak/dependencies-sources.json ./distribution/flatpak/gradle-sources-app-core.json
 cp ./modules/app-desktop/build/flatpak/dependencies-sources.json ./distribution/flatpak/gradle-sources-app-desktop.json
+cp ./modules/app-ui/build/flatpak/dependencies-sources.json ./distribution/flatpak/gradle-sources-app-ui.json
 cp ./modules/cli/build/flatpak/dependencies-sources.json ./distribution/flatpak/gradle-sources-cli.json
 cp ./modules/files/build/flatpak/dependencies-sources.json ./distribution/flatpak/gradle-sources-files.json