+++ modules/app-android/src/androidMain/kotlin/org/archivekeep/app/android/AndroidEnvironment.kt
@@ -1,5 +1,6 @@
 package org.archivekeep.app.android
 
+import android.content.Context
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.serializer
@@ -7,7 +8,9 @@ import org.archivekeep.app.core.persistence.credentials.Credentials
 import org.archivekeep.app.core.persistence.credentials.CredentialsInProtectedDataStore
 import org.archivekeep.app.core.persistence.credentials.CredentialsStore
 import org.archivekeep.app.core.persistence.credentials.JoseStorage
+import org.archivekeep.app.core.persistence.drivers.filesystem.AndroidFileStores
 import org.archivekeep.app.core.persistence.drivers.filesystem.FileStores
+import org.archivekeep.app.core.persistence.drivers.filesystem.FileSystemStorageDriver
 import org.archivekeep.app.core.persistence.drivers.grpc.GRPCStorageDriver
 import org.archivekeep.app.core.persistence.platform.Environment
 import org.archivekeep.app.core.persistence.registry.PreferenceDataStoreRegistryData
@@ -15,10 +18,12 @@ import org.archivekeep.app.core.persistence.repository.MemorizedRepositoryIndexR
 import org.archivekeep.app.core.persistence.repository.MemorizedRepositoryMetadataRepositoryInDataStore
 
 class AndroidEnvironment(
+    context: Context,
     val scope: CoroutineScope,
     paths: AndroidEnvironmentPaths,
-    override val fileStores: FileStores = FileStores(scope),
 ) : Environment {
+    override val fileStores: FileStores = AndroidFileStores(context, scope)
+
     override val registry: PreferenceDataStoreRegistryData = PreferenceDataStoreRegistryData(scope, paths.getRegistryDatastoreFile())
 
     override val walletDataStore =
@@ -35,8 +40,7 @@ class AndroidEnvironment(
 
     override val storageDrivers =
         mapOf(
-            // TODO: this will be less straightforward - Android has own filesystem API -> Storage Access Framework
-            // "filesystem" to FileSystemStorageDriver(scope, fileStores),
+            "filesystem" to FileSystemStorageDriver(scope, fileStores),
             "grpc" to GRPCStorageDriver(credentialsStore),
         )
 }