+++ modules/app-android/src/androidMain/kotlin/org/archivekeep/app/android/MainApplication.kt
@@ -18,6 +18,7 @@ class MainApplication : Application() {
         services =
             ApplicationServices(scope) { scope ->
                 AndroidEnvironment(
+                    applicationContext,
                     scope + Dispatchers.Default,
                     AndroidEnvironmentPaths(filesDir),
                 )