+++ modules/app-ui/src/androidMain/kotlin/org/archivekeep/app/desktop/domain/services/RepositoryOpenService.android.kt
@@ -1,0 +1,5 @@
+package org.archivekeep.app.desktop.domain.services
+
+import org.archivekeep.app.desktop.domain.wiring.ApplicationServices
+
+actual fun createRepositoryOpenService(applicationServices: ApplicationServices): RepositoryOpenService = AndroidRepositoryOpenService()