+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Changed
 
 - refactor: extract UI to app-ui module
+- build: limit collection of Gradle dependencies for Flatpak to sourceSets needed for its build
 
 ## 0.3.3 - 2025-03-31